refactor: Part 1
