+++ tsconfig.test.json
@@ -1,0 +1,23 @@
+{
+    "compilerOptions": {
+        "esModuleInterop": true,
+        "target": "ESNext",
+        "module": "ESNext",
+        "moduleResolution": "Node",
+        "noImplicitThis": true,
+        "skipLibCheck": true,
+        "resolveJsonModule": true,
+        "emitDeclarationOnly": true,
+        "declaration": true,
+        "declarationMap": true,
+        "declarationDir": "dist/declaration",
+        // always use import type for type only imports
+        "verbatimModuleSyntax": true,
+        "strict": true
+    },
+    "include": [
+        "src/**/*",
+        "scripts/**/*",
+        "tests/**/*"
+    ]
+}
