+++ vitest.config.ts
@@ -8,6 +8,7 @@ export default defineConfig({
         fileParallelism: false,
         typecheck: {
             ignoreSourceErrors: true,
+            tsconfig: "./tsconfig.test.json",
         },
     },
 });
